+++ lib/custom_snack_bar.dart
@@ -8,11 +8,19 @@ class CustomSnackBar extends StatefulWidget {
   final Widget icon;
   final Color backgroundColor;
   final TextStyle textStyle;
+  final int maxLines;
   final int iconRotationAngle;
+  final List<BoxShadow> boxShadow;
+  final BorderRadius borderRadius;
+  final double iconPositionTop;
+  final double iconPositionLeft;
+  final EdgeInsetsGeometry messagePadding;
+  final double textScaleFactor;
 
   const CustomSnackBar.success({
     Key? key,
     required this.message,
+    this.messagePadding = const EdgeInsets.symmetric(horizontal: 24),
     this.icon = const Icon(
       Icons.sentiment_very_satisfied,
       color: const Color(0x15000000),
@@ -23,13 +31,20 @@ class CustomSnackBar extends StatefulWidget {
       fontSize: 16,
       color: Colors.white,
     ),
+    this.maxLines = 2,
     this.iconRotationAngle = 32,
+    this.iconPositionTop = -10,
+    this.iconPositionLeft = -8,
     this.backgroundColor = const Color(0xff00E676),
+    this.boxShadow = kDefaultBoxShadow,
+    this.borderRadius = kDefaultBorderRadius,
+    this.textScaleFactor = 1.0,
   });
 
   const CustomSnackBar.info({
     Key? key,
     required this.message,
+    this.messagePadding = const EdgeInsets.symmetric(horizontal: 24),
     this.icon = const Icon(
       Icons.sentiment_neutral,
       color: const Color(0x15000000),
@@ -40,13 +55,20 @@ class CustomSnackBar extends StatefulWidget {
       fontSize: 16,
       color: Colors.white,
     ),
+    this.maxLines = 2,
     this.iconRotationAngle = 32,
+    this.iconPositionTop = -10,
+    this.iconPositionLeft = -8,
     this.backgroundColor = const Color(0xff2196F3),
+    this.boxShadow = kDefaultBoxShadow,
+    this.borderRadius = kDefaultBorderRadius,
+    this.textScaleFactor = 1.0,
   });
 
   const CustomSnackBar.error({
     Key? key,
     required this.message,
+    this.messagePadding = const EdgeInsets.symmetric(horizontal: 24),
     this.icon = const Icon(
       Icons.error_outline,
       color: const Color(0x15000000),
@@ -57,8 +79,14 @@ class CustomSnackBar extends StatefulWidget {
       fontSize: 16,
       color: Colors.white,
     ),
+    this.maxLines = 2,
     this.iconRotationAngle = 32,
+    this.iconPositionTop = -10,
+    this.iconPositionLeft = -8,
     this.backgroundColor = const Color(0xffff5252),
+    this.boxShadow = kDefaultBoxShadow,
+    this.borderRadius = kDefaultBorderRadius,
+    this.textScaleFactor = 1.0,
   });
 
   @override
@@ -70,38 +98,30 @@ class _CustomSnackBarState extends State<CustomSnackBar> {
   Widget build(BuildContext context) {
     final theme = Theme.of(context);
     return Container(
+      clipBehavior: Clip.hardEdge,
       height: 80,
       decoration: BoxDecoration(
         color: widget.backgroundColor,
-        borderRadius: BorderRadius.all(Radius.circular(12)),
-        boxShadow: [
-          BoxShadow(
-            color: Colors.black26,
-            offset: Offset(0.0, 8.0),
-            spreadRadius: 1,
-            blurRadius: 30,
-          ),
-        ],
+        borderRadius: widget.borderRadius,
+        boxShadow: widget.boxShadow,
       ),
       width: double.infinity,
       child: Stack(
         children: [
           Positioned(
-            top: -10,
-            left: -8,
-            child: ClipRRect(
-              child: Container(
-                height: 95,
-                child: Transform.rotate(
-                  angle: widget.iconRotationAngle * pi / 180,
-                  child: widget.icon,
-                ),
+            top: widget.iconPositionTop,
+            left: widget.iconPositionLeft,
+            child: Container(
+              height: 95,
+              child: Transform.rotate(
+                angle: widget.iconRotationAngle * pi / 180,
+                child: widget.icon,
               ),
             ),
           ),
           Center(
             child: Padding(
-              padding: const EdgeInsets.symmetric(horizontal: 24),
+              padding: widget.messagePadding,
               child: Text(
                 widget.message,
                 style: theme.textTheme.bodyText2?.merge(
@@ -109,7 +129,8 @@ class _CustomSnackBarState extends State<CustomSnackBar> {
                 ),
                 textAlign: TextAlign.center,
                 overflow: TextOverflow.ellipsis,
-                maxLines: 2,
+                maxLines: widget.maxLines,
+                textScaleFactor: widget.textScaleFactor,
               ),
             ),
           ),
@@ -118,3 +139,14 @@ class _CustomSnackBarState extends State<CustomSnackBar> {
     );
   }
 }
+
+const kDefaultBoxShadow = const [
+  BoxShadow(
+    color: Colors.black26,
+    offset: Offset(0.0, 8.0),
+    spreadRadius: 1,
+    blurRadius: 30,
+  ),
+];
+
+const kDefaultBorderRadius = const BorderRadius.all(Radius.circular(12));