+++ lib/top_snack_bar.dart
@@ -1,36 +1,52 @@
 import 'package:flutter/material.dart';
 import 'package:top_snackbar_flutter/tap_bounce_container.dart';
 
+typedef ControllerCallback = void Function(AnimationController);
+
+OverlayEntry? _previousEntry;
+
 /// Displays a widget that will be passed to [child] parameter above the current
 /// contents of the app, with transition animation
 ///
 /// The [child] argument is used to pass widget that you want to show
 ///
-/// The [showOutAnimationDuration] argument is used to specify duration of
+/// The [animationDuration] argument is used to specify duration of
 /// enter transition
 ///
-/// The [hideOutAnimationDuration] argument is used to specify duration of
+/// The [reverseAnimationDuration] argument is used to specify duration of
 /// exit transition
 ///
 /// The [displayDuration] argument is used to specify duration displaying
-///
-/// The [additionalTopPadding] argument is used to specify amount of top
-/// padding that will be added for SafeArea values
 ///
 /// The [onTap] callback of [TopSnackBar]
 ///
 /// The [overlayState] argument is used to add specific overlay state.
 /// If you will not pass it, it will try to get the current overlay state from
 /// passed [BuildContext]
+///
+/// The [persistent] argument is used to make snack bar persistent, so
+/// [displayDuration] will be ignored. Default is false.
+///
+/// The [onAnimationControllerInit] callback is called on internal
+/// [AnimationController] has been initialized.
+///
+/// The [padding] argument is used to specify amount of outer padding
+///
+/// [curve] and [reverseCurve] arguments are used to specify curves
+/// for in and out animations respectively
 void showTopSnackBar(
   BuildContext context,
   Widget child, {
-  Duration showOutAnimationDuration = const Duration(milliseconds: 1200),
-  Duration hideOutAnimationDuration = const Duration(milliseconds: 550),
+  Duration animationDuration = const Duration(milliseconds: 1200),
+  Duration reverseAnimationDuration = const Duration(milliseconds: 550),
   Duration displayDuration = const Duration(milliseconds: 3000),
-  double additionalTopPadding = 16.0,
   VoidCallback? onTap,
   OverlayState? overlayState,
+  bool persistent = false,
+  ControllerCallback? onAnimationControllerInit,
+  EdgeInsets padding = const EdgeInsets.all(16),
+  Curve curve = Curves.elasticOut,
+  Curve reverseCurve = Curves.linearToEaseOut,
 }) async {
   overlayState ??= Overlay.of(context);
   late OverlayEntry overlayEntry;
@@ -38,62 +54,90 @@ void showTopSnackBar(
     builder: (context) {
       return TopSnackBar(
         child: child,
-        onDismissed: () => overlayEntry.remove(),
-        showOutAnimationDuration: showOutAnimationDuration,
-        hideOutAnimationDuration: hideOutAnimationDuration,
+        onDismissed: () {
+          if (overlayEntry.mounted && _previousEntry == overlayEntry) {
+            overlayEntry.remove();
+            _previousEntry = null;
+          }
+        },
+        animationDuration: animationDuration,
+        reverseAnimationDuration: reverseAnimationDuration,
         displayDuration: displayDuration,
-        additionalTopPadding: additionalTopPadding,
         onTap: onTap,
+        persistent: persistent,
+        onAnimationControllerInit: onAnimationControllerInit,
+        padding: padding,
+        curve: curve,
+        reverseCurve: reverseCurve,
       );
     },
   );
 
+  if (_previousEntry != null && _previousEntry!.mounted) {
+    _previousEntry?.remove();
+  }
   overlayState?.insert(overlayEntry);
+  _previousEntry = overlayEntry;
 }
 
 /// Widget that controls all animations
 class TopSnackBar extends StatefulWidget {
   final Widget child;
   final VoidCallback onDismissed;
-  final showOutAnimationDuration;
-  final hideOutAnimationDuration;
-  final displayDuration;
-  final additionalTopPadding;
+  final Duration animationDuration;
+  final Duration reverseAnimationDuration;
+  final Duration displayDuration;
   final VoidCallback? onTap;
+  final ControllerCallback? onAnimationControllerInit;
+  final bool persistent;
+  final EdgeInsets padding;
+  final Curve curve;
+  final Curve reverseCurve;
 
   TopSnackBar({
     Key? key,
     required this.child,
     required this.onDismissed,
-    required this.showOutAnimationDuration,
-    required this.hideOutAnimationDuration,
+    required this.animationDuration,
+    required this.reverseAnimationDuration,
     required this.displayDuration,
-    required this.additionalTopPadding,
     this.onTap,
+    this.persistent = false,
+    this.onAnimationControllerInit,
+    required this.padding,
+    required this.curve,
+    required this.reverseCurve,
   }) : super(key: key);
 
   @override
   _TopSnackBarState createState() => _TopSnackBarState();
 }
 
-class _TopSnackBarState extends State<TopSnackBar> with SingleTickerProviderStateMixin {
+class _TopSnackBarState extends State<TopSnackBar>
+    with SingleTickerProviderStateMixin {
   late Animation offsetAnimation;
   late AnimationController animationController;
-  double? topPosition;
 
   @override
   void initState() {
-    topPosition = widget.additionalTopPadding;
     _setupAndStartAnimation();
     super.initState();
+  }
+
+  @override
+  void dispose() {
+    animationController.dispose();
+    super.dispose();
   }
 
   void _setupAndStartAnimation() async {
     animationController = AnimationController(
       vsync: this,
-      duration: widget.showOutAnimationDuration,
-      reverseDuration: widget.hideOutAnimationDuration,
+      duration: widget.animationDuration,
+      reverseDuration: widget.reverseAnimationDuration,
     );
+
+    widget.onAnimationControllerInit?.call(animationController);
 
     Tween<Offset> offsetTween = Tween<Offset>(
       begin: Offset(0.0, -1.0),
@@ -103,43 +147,52 @@ class _TopSnackBarState extends State<TopSnackBar> with SingleTickerProviderStat
     offsetAnimation = offsetTween.animate(
       CurvedAnimation(
         parent: animationController,
-        curve: Curves.elasticOut,
-        reverseCurve: Curves.linearToEaseOut,
+        curve: widget.curve,
+        reverseCurve: widget.reverseCurve,
       ),
     )..addStatusListener((status) async {
         if (status == AnimationStatus.completed) {
-          await Future.delayed(widget.displayDuration);
-          animationController.reverse();
-          if (mounted) {
-            setState(() {
-              topPosition = 0;
-            });
+          if (!widget.persistent) {
+            await Future.delayed(widget.displayDuration);
+            _dismiss();
           }
         }
 
         if (status == AnimationStatus.dismissed) {
-          widget.onDismissed.call();
+          if (mounted) {
+            widget.onDismissed.call();
+          }
         }
       });
 
-    animationController.forward();
+    if (mounted) {
+      animationController.forward();
+    }
+  }
+
+  void _dismiss() {
+    if (mounted) {
+      animationController.reverse();
+    }
   }
 
   @override
   Widget build(BuildContext context) {
-    return AnimatedPositioned(
-      duration: widget.hideOutAnimationDuration * 1.5,
-      curve: Curves.linearToEaseOut,
-      top: topPosition,
-      left: 16,
-      right: 16,
+    return Positioned(
+      top: widget.padding.top,
+      left: widget.padding.left,
+      right: widget.padding.right,
       child: SlideTransition(
         position: offsetAnimation as Animation<Offset>,
         child: SafeArea(
           child: TapBounceContainer(
             onTap: () {
-              widget.onTap?.call();
-              animationController.reverse();
+              if (mounted) {
+                widget.onTap?.call();
+                if (!widget.persistent) {
+                  animationController.reverse();
+                }
+              }
             },
             child: widget.child,
           ),